Common template engine.
