+++ ql/pricingengines/lookback/all.hpp
@@ -5,8 +5,5 @@
 #include <ql/pricingengines/lookback/analyticcontinuousfloatinglookback.hpp>
 #include <ql/pricingengines/lookback/analyticcontinuouspartialfixedlookback.hpp>
 #include <ql/pricingengines/lookback/analyticcontinuouspartialfloatinglookback.hpp>
-#include <ql/pricingengines/lookback/mclookbackpartialfixedengine.hpp>
-#include <ql/pricingengines/lookback/mclookbackfixedengine.hpp>
-#include <ql/pricingengines/lookback/mclookbackfloatingengine.hpp>
-#include <ql/pricingengines/lookback/mclookbackpartialfloatingengine.hpp>
+#include <ql/pricingengines/lookback/mclookbackengine.hpp>
